requirements_dev.txt
-r requirements.txt
django-debug-panel==0.8.3
django-debug-toolbar==1.4